cart_service/requirements.txt
fastapi
uvicorn
redis
httpx
pydantic
pika
python-jose[cryptography]